Added options and recent helper
Switched from using outdated UnityASL to the general asl-help. Script allows you to start on making a new file, split on results screens (not yet Utopia Shelter) and resets if you delete the same file you just played on. Minor bugfix to timing outside of the level: the game doesn't reset IsPaused when exiting a level so the variable stays true in the overworld, have to make exceptions for specific scenes.